debugging

+++ models/finetune.py
@@ -57,7 +57,7 @@ class Finetune(CLIPInterface):
         enable_pos_emb: bool = False,
     ):
         super().__init__(
-            clip_model.float(),
+            clip_model.half(),
             config,
             token_ids,
             soft_embeddings=torch.zeros([len(attributes)]),
@@ -70,8 +70,8 @@ class Finetune(CLIPInterface):
         self.attributes = attributes
         self.objects = objects
         self.prompt_template = prompt_template
-        self.dtype = torch.float32
-        self.text_encoder = self.text_encoder.float()
+        self.dtype = torch.float16
+        self.text_encoder = self.text_encoder.half()
 
     def construct_token_tensors(self, pair_idx):
         attr_idx, obj_idx = pair_idx[:, 0], pair_idx[:, 1]
